requirements.txt
Django==1.11.7
djangorestframework==3.7.3
pkg-resources==0.0.0
psycopg2==2.7.3.2
python-decouple==3.1
pytz==2017.3
